[NO ISSUE][STO] Column cloud refactoring
- user model changes: no
- storage format changes: no
- interface changes: no

Details:
Refactor common code and some cleanups in cloud
column read path.

Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/18382
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>

+++ hyracks-fullstack/hyracks/hyracks-storage-am-lsm-btree-column/src/main/java/org/apache/hyracks/storage/am/lsm/btree/column/impls/btree/ColumnBTreeReadLeafFrame.java
@@ -90,7 +90,7 @@ public final class ColumnBTreeReadLeafFrame extends AbstractColumnBTreeLeafFrame
         return buf.getInt(NEXT_LEAF_OFFSET);
     }
 
-    public long getMegaLeafNodeLengthInBytes() {
+    public int getMegaLeafNodeLengthInBytes() {
         return buf.getInt(MEGA_LEAF_NODE_LENGTH);
     }
